ci: add pre-commit hooks
Found github:cachix/pre-commit-hooks.nix which provides pre-commit hooks
in the Nix style. It needs an option to be initialized first, so the
hook is written to the git hooks folder.

This is simple done passing a shell hook to a nix develop shell as
described in the README of the project.