feat: add tmux-helper-script ref-main
The helper script checks if a tmux main session is present, then
probably starts one and switches to it.

Injection using Nix overlay and `writeShellApplication`.

The `writeShellApplication` has the benefit that it checks the concatenated
shell script with shellchecker, while `writeShellScriptBin` not performs
such a check.

Probably it's a good idea to disable the shellcheck SC2148 for external
scripts when there's no first interpreter line present.