refactor: rewrite and document system and pkgs functions
Take inspiration from Misterio77 nix-config and integrate his pkgsFor
and forEachSystem functions into my nix config. To align it for my
needs, rename the existing forAllSystems functions to forEachSystem
and rename the Misterios77 forEachSystem to pkgsForEachSystem, as
it's intention is to give access to pkgs for each current system.

Align also the names my pkgs function to pkgsFor and unstablePkgsFor.

In a later iteration I will integrate the unstable pkgs access using
an overlay.