build: refactor and update pre-commit hooks
Update flake input for pre-commit-hooks, as the repo was renamed to git-hooks.

Refactoring the pre-commit-hooks configuration, following the example
provided in the README to use checks for that. Instead of keeping the
checks in `flake.nix` directly, start to manage the checks in the
checks dir.

Rename the original buildInputs attribute of mkShell as
nativeBuildInputs should be used for commands that need to run
at build time like cmake or shell hooks, according to a discourse
thread.

So that would also be the case for the pkgs required by the
pre-commit-checks, but I keep the config as given in the README
first.